+++ mayan/apps/sources/source_backend_actions/mixins/document_language_mixins.py
@@ -7,7 +7,25 @@ from ..interfaces import (
 from .arguments import argument_document_language
 
 
-class SourceBackendActionMixinLanguageInteractive:
+class SourceBackendActionMixinLanguageBase:
+    def get_document_task_kwargs(self, language=None, **kwargs):
+        result = super().get_document_task_kwargs(**kwargs)
+
+        result['language'] = language
+
+        return result
+
+    def get_task_kwargs(self, language=None, **kwargs):
+        result = super().get_task_kwargs(**kwargs)
+
+        result['action_interface_kwargs']['language'] = language
+
+        return result
+
+
+class SourceBackendActionMixinLanguageInteractive(
+    SourceBackendActionMixinLanguageBase
+):
     class Interface:
         class Model(SourceBackendActionInterface):
             class Argument:
@@ -47,16 +65,15 @@ class SourceBackendActionMixinLanguageInteractive:
                     'language', None
                 )
 
-    def get_document_task_kwargs(self, language=None, **kwargs):
-        result = super().get_document_task_kwargs(**kwargs)
 
-        result['language'] = language
+class SourceBackendActionMixinLanguageInteractiveNot(
+    SourceBackendActionMixinLanguageBase
+):
+    class Interface:
+        class Task(SourceBackendActionInterfaceTask):
+            def process_interface_context(self):
+                super().process_interface_context()
 
-        return result
+                source_backend_instance = self.action.source.get_backend_instance()
 
-    def get_task_kwargs(self, language=None, **kwargs):
-        result = super().get_task_kwargs(**kwargs)
-
-        result['action_interface_kwargs']['language'] = language
-
-        return result
+                self.action_kwargs['language'] = source_backend_instance.get_document_language()