+++ HISTORY.rst
@@ -64,6 +64,10 @@
   - Support automatic nested field ordering.
   - Add testing for field ordering.
 
+- Support settings the document language for periodic sources. Thanks to
+  forum user @jeverling for the request.
+  https://forum.mayan-edms.com/t/feature-suggestion-allow-specifying-default-language-for-sources/762
+
 4.6.2 (2024-03-04)
 ==================
 - Add clamav to the makefile ``setup-dev-operating-system-packages`` target.